refactor: RequestLineParser와 HttpRequest 클래스 통합
