+++ src/test/java/domain/RequestLineTest.java
@@ -9,45 +9,38 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static org.assertj.core.api.Assertions.*;
 
 @DisplayName("RequestLineParser 단위 테스트")
-public class RequestLineParserTest {
-    private RequestLineParser requestLineParser;
-
-    @BeforeEach
-    void setUp() {
-        requestLineParser = new RequestLineParser();
-    }
-
+public class RequestLineTest {
     @DisplayName("GET RequestLine을 파싱하여 HttpRequest를 얻는다.")
     @Test
     void parseForGet() {
         // when
-        HttpRequest httpRequest = requestLineParser.parse("GET /users HTTP/1.1");
+        RequestLine requestLine = new RequestLine("GET /users HTTP/1.1");
 
         // then
-        assertThat(httpRequest).isEqualTo(
-                new HttpRequest(HttpMethod.GET, new HttpPath("/users"), new HttpProtocol("HTTP/1.1")));
+        assertThat(requestLine).isEqualTo(
+                new RequestLine(HttpMethod.GET, new HttpPath("/users"), new HttpProtocol("HTTP/1.1")));
     }
 
     @DisplayName("POST RequestLine을 파싱하여 HttpRequest를 얻는다.")
     @Test
     void parseForPost() {
         // when
-        HttpRequest httpRequest = requestLineParser.parse("POST /users HTTP/1.1");
+        RequestLine requestLine = new RequestLine("POST /users HTTP/1.1");
 
         // then
-        assertThat(httpRequest).isEqualTo(
-                new HttpRequest(HttpMethod.POST, new HttpPath("/users"), new HttpProtocol("HTTP/1.1")));
+        assertThat(requestLine).isEqualTo(
+                new RequestLine(HttpMethod.POST, new HttpPath("/users"), new HttpProtocol("HTTP/1.1")));
     }
 
     @DisplayName("Query String이 있는 RequestLine을 파싱하여 HttpRequest를 얻는다.")
     @Test
     void parseWithQueryString() {
         // when
-        HttpRequest httpRequest = requestLineParser.parse("GET /users?userId=javajigi&password=password&name=JaeSung HTTP/1.1");
+        RequestLine requestLine = new RequestLine("GET /users?userId=javajigi&password=password&name=JaeSung HTTP/1.1");
 
         // then
-        assertThat(httpRequest).isEqualTo(
-                new HttpRequest(HttpMethod.GET, new HttpPath("/users?userId=javajigi&password=password&name=JaeSung"), new HttpProtocol("HTTP/1.1")));
+        assertThat(requestLine).isEqualTo(
+                new RequestLine(HttpMethod.GET, new HttpPath("/users?userId=javajigi&password=password&name=JaeSung"), new HttpProtocol("HTTP/1.1")));
     }
 
     @DisplayName("HTTP 요청 형식이 맞지 않으면 예외를 발생한다.")
@@ -57,8 +50,8 @@ public class RequestLineParserTest {
             "GET /users HTTP/1.1 HTTP/1.1",
     })
     void parseException(String requestLine) {
-        assertThatThrownBy(() -> requestLineParser.parse(requestLine))
+        assertThatThrownBy(() -> new RequestLine(requestLine))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(RequestLineParser.VALIDATION_MESSAGE);
+                .hasMessage(RequestLine.VALIDATION_MESSAGE);
     }
 }
